Move Handshake types to Framework
- Move HandshakeOptions to src/Framework/Backend
- Move HandshakeStatus to src/Framework/Backend
- Move HandshakeResult to src/Framework/Backend
- Move HandshakeComponents to src/Framework/Backend
- Move Handshake to src/Framework/Backend
- Move ServerNodeHandshake to src/Framework/Backend
